Fix readability and memory leak
